+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## Added
 
+* `plot_pvalue_barplot`: horizontal barplot of pvalues
 * `plot_dotmap`: plot heatmap of effect sizes (dots) with cells shaded by pvalue
+    + option to add Fisher's combination pvalue barplot on the right
 
 ## Changed